new file structure: scr/devices & scr/ultrasound
21/12/2022
