+++ drift_check_Pico.py
@@ -8,15 +8,10 @@ Python version: Python 3.8
 """
 import numpy as np
 import matplotlib.pyplot as plt
-import ctypes
 import time
 
-# import Pico5000alib as plib
-# import US_Functions as USF
-# import US_GenCode as USGC
-# import US_Graphics as USG
-import ultrasound as US
-import pico
+import src.ultrasound as US
+from src.devices import pico5000a
 
 #%% Parameters
 num_bits = 15               # Number of bits to use (8, 12, 14, 15 or 16) - int
@@ -142,31 +137,26 @@ ARBITRARY_SIGNAL_GENERATOR_DICT = {
 
 #%% Initial check
 # Find out device model (5000a)
-# plib.check_drivers()
+# pico5000a.check_drivers()
 
 
-#%% Start
-# Create chandle and status ready for use
-chandle = ctypes.c_int16()
-status = {}
-
-# Start pico
-pico.start_pico5000a(chandle, status, num_bits)
+#%% Start pico
+pico = pico5000a.Pico(num_bits)
 
 #%% Setup
 # Set up channel A
-pico.setup_channel(chandle, status, 'A', coupling_A, voltage_range_A, offset_A, enabled_A)
+pico.setup_channel('A', coupling_A, voltage_range_A, offset_A, enabled_A)
 
 # Set up channel B
-pico.setup_channel(chandle, status, 'B', coupling_B, voltage_range_B, offset_B, enabled_B)
+pico.setup_channel('B', coupling_B, voltage_range_B, offset_B, enabled_B)
 
 # Set up simple trigger
 voltage_range = voltage_range_B if triggerChannel=='B' else voltage_range_A
-pico.set_simpleTrigger(chandle, status, enabled_trigger, triggerChannel, voltage_range, triggerThreshold, direction, delay, auto_Trigger)
+pico.set_simpleTrigger(enabled_trigger, triggerChannel, voltage_range, triggerThreshold, direction, delay, auto_Trigger)
 
 #%% Get timebase
 # Get timebase
-timebase, timeIntervalns, maxSamples = pico.get_timebase(chandle, status, Fs, preTriggerSamples + postTriggerSamples, segmentIndex=0)
+timebase, timeIntervalns, maxSamples = pico.get_timebase(Fs, preTriggerSamples + postTriggerSamples, segmentIndex=0)
 Real_Fs = 1e9/(2**timebase) # Hz
 
 if generate_arbitrary_signal and ARBITRARY_SIGNAL_GENERATOR_DICT['startFrequency'] == ARBITRARY_SIGNAL_GENERATOR_DICT['stopFrequency']:
@@ -178,11 +168,11 @@ if generate_arbitrary_signal and ARBITRARY_SIGNAL_GENERATOR_DICT['startFrequency
 
 #%% Generate signal
 if generate_arbitrary_signal:
-    pico.generate_arbitrary_signal(chandle, status, **ARBITRARY_SIGNAL_GENERATOR_DICT)
+    pico.generate_arbitrary_signal(**ARBITRARY_SIGNAL_GENERATOR_DICT)
     triggertype = ARBITRARY_SIGNAL_GENERATOR_DICT['triggertype']
     triggerSource = ARBITRARY_SIGNAL_GENERATOR_DICT['triggerSource']
 elif generate_builtin_signal:
-    pico.generate_builtin_signal(chandle, status, **BUILTIN_SIGNAL_GENERATOR_DICT)
+    pico.generate_builtin_signal(**BUILTIN_SIGNAL_GENERATOR_DICT)
     triggertype = BUILTIN_SIGNAL_GENERATOR_DICT['triggertype']
     triggerSource = BUILTIN_SIGNAL_GENERATOR_DICT['triggerSource']
 
@@ -196,20 +186,20 @@ ToF = np.zeros(N_acqs)
 for i in range(N_acqs):
     # Run rapid capture of nSegments
     # BUFFERS_DICT, cmaxSamples, triggerTimeOffsets, triggerTimeOffsetUnits, time_indisposed, triggerInfo = pico.rapid_capture(
-    #     chandle, status, channels, (preTriggerSamples, postTriggerSamples), timebase,
+    #     channels, (preTriggerSamples, postTriggerSamples), timebase,
     #     nSegments, trigger_sigGen, triggertype, gate_time,
     #     downsampling=(downsampling_ratio_mode, downsampling_ratio))
     BUFFERS_DICT, cmaxSamples, triggerTimeOffset, triggerTimeOffsetUnits, time_indisposed = pico.capture(
-            chandle, status, channels, (preTriggerSamples, postTriggerSamples), timebase, 
+            channels, (preTriggerSamples, postTriggerSamples), timebase, 
             trigger_sigGen, triggertype, gate_time, downsampling=(downsampling_ratio_mode, downsampling_ratio), segment_index=0)
-    means = pico.get_data_from_buffersdict(chandle, status, voltage_range_A, voltage_range_B, BUFFERS_DICT)[4]
+    means = pico.get_data_from_buffersdict(voltage_range_A, voltage_range_B, BUFFERS_DICT)[4]
     
     start_time = time.time()
     
     # Create time data
     t = np.linspace(0, (cmaxSamples - 1) * timeIntervalns, cmaxSamples)
     
-    means = pico.get_data_from_buffersdict(chandle, status, voltage_range_A, voltage_range_B, BUFFERS_DICT)[4]
+    means = pico.get_data_from_buffersdict(voltage_range_A, voltage_range_B, BUFFERS_DICT)[4]
     
     TT = means[0]
     if i==0: TT0 = means[0]
@@ -218,8 +208,8 @@ for i in range(N_acqs):
     US.movefig(location='southeast')
     axs[0].plot(t*1e-3, TT0, lw=2)
     axs[1].plot(t*1e-3, TT, lw=2)
-    axs[0].set_ylim([-pico.str2V(voltage_range_A)*1e3, pico.str2V(voltage_range_A)*1e3])
-    axs[1].set_ylim([-pico.str2V(voltage_range_A)*1e3, pico.str2V(voltage_range_A)*1e3])
+    axs[0].set_ylim([-pico5000a.str2V(voltage_range_A)*1e3, pico5000a.str2V(voltage_range_A)*1e3])
+    axs[1].set_ylim([-pico5000a.str2V(voltage_range_A)*1e3, pico5000a.str2V(voltage_range_A)*1e3])
     axs[1].set_xlabel('Time (us)')
     axs[0].set_ylabel('Voltage (mV)')
     axs[1].set_ylabel('Voltage (mV)')
@@ -253,9 +243,9 @@ for i in range(N_acqs):
     time.sleep(time_to_wait)
     
 # Stop the scope
-pico.stop_pico5000a(chandle, status)
+pico.stop()
 
 
 #%% Close
 # Close the unit
-pico.close_pico5000a(chandle, status)
+pico.close()
