+++ src/__init__.py
@@ -1,0 +1,8 @@
+# -*- coding: utf-8 -*-
+"""
+Created on Wed Dec 21 13:21:30 2022
+Python version: Python 3.8
+
+@author: Arnau Busqué Nadal <arnau.busque@goumh.umh.es>
+
+"""
